[pattern-recognition] Wrapped functions in enum
